WIP: crash comiccrawler.core
