+++ comiccrawler/core.py
@@ -1,0 +1,5 @@
+#! python3
+# pylint: disable=unused-import
+# for compatibility
+from .grabber import grabhtml 
+from .episode import Episode
